アニメーションバグ修正＆コントローラ操作を修正

+++ Project1/Project1/ObjCloud.h
@@ -51,7 +51,7 @@ private:
 	float m_con_x;
 	float m_con_y;
 
-	int   m_ani_time;   //アニメーションフレーム動作間隔
+	float   m_ani_time;   //アニメーションフレーム動作間隔
 	int   m_ani_frame;  //描画フレーム
 
 	float m_speed_power;  //スピードパワー
